# Conflicts:
#	app/src/main/java/com/barisetech/www/workmanage/base/BaseConstant.java
#	app/src/main/java/com/barisetech/www/workmanage/view/fragment/MapFragment.java
#	app/src/main/java/com/barisetech/www/workmanage/view/fragment/Messagefragment.java
#	app/src/main/java/com/barisetech/www/workmanage/view/fragment/PipeCollectionFragment.java

+++ app/src/main/java/com/barisetech/www/workmanage/view/fragment/PipeModifyFragment.java
@@ -219,7 +219,7 @@ public class PipeModifyFragment extends BaseFragment {
 //                curPipeInfo.PipeCollectID = null;
 //            }
 
-            curPipeInfo.Length = Integer.valueOf(length);
+            curPipeInfo.Length = Double.valueOf(length);
             curPipeInfo.PipeMaterial = materail;
 //            curPipeInfo.Company = company;
             curPipeInfo.Speed = Integer.valueOf(speed);
